WIP: query status directly from DB
working on querying sipity entities directly from the database. once we have a
list of ids, we pull their documents from Solr.

with those in hand, we probably want to get data into a presenter next and add
the workflow states in alongside. the presenter should be able to determine if
the workflow state is `#published?` for some definition thereof within the
workflow.

aside: is it a good assumption that workflows have a `#published` status? the
workflow use cases seem to extend far beyond publication workflows. do we like
the tabs?